+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Changes:
 * Save log to a file and show its path on settings(closes #7)
 * Save window dimensions(closes #14)
 * Add theme selection and show black text when white theme is selected(closes #18)
+* Implement color switcher feature(closes #22)